+++ src/Cart/CartRestoreException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Cart;
+
+use Exception;
+
+class CartRestoreException extends Exception {}